Added newly trained custom model based on ssd_mobilenet_v2_320x320 and fixed bug in partition dataset

+++ object_detection/scripts/preprocessing/partition_dataset.py
@@ -35,7 +35,7 @@ def iterate_dir(source, dest, ratio, copy_xml):
               if re.search(r'([a-zA-Z0-9\s_\\.\-\(\):])+(?i)(.jpg|.jpeg|.png)$', f)]
 
     num_images = len(images)
-    num_test_images = math.ceil(ratio*num_images)
+    num_test_images = int(math.ceil(ratio*num_images))
 
     for i in range(num_test_images):
         idx = random.randint(0, len(images)-1)
